requirements.txt
flask
flask-sqlalchemy
flask-login
flask-bcrypt 
python-dateutil